Refactor test fakes into dedicated testing modules
