+++ core/camera/testing/src/main/java/com/google/jetpackcamera/core/camera/testing/FakeCameraSystem.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.core.camera.test
+package com.google.jetpackcamera.core.camera.testing
 
 import android.annotation.SuppressLint
 import android.content.ContentResolver
