+++ data/settings/testing/src/main/java/com/google/jetpackcamera/settings/testing/FakeDataStoreModule.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.jetpackcamera.settings.test
+package com.google.jetpackcamera.settings.testing
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.core.DataStoreFactory
