+++ feature/preview/build.gradle.kts
@@ -141,7 +141,10 @@ dependencies {
     implementation(project(":data:media"))
     implementation(project(":data:settings"))
     implementation(project(":core:model"))
-    testImplementation(project(":core:common"))
+    testImplementation(project(":core:camera:testing"))
+    testImplementation(project(":data:settings:testing"))
+    testImplementation(project(":data:media:testing"))
+    testImplementation(project(":core:common:testing"))
     implementation(project(":ui:components:capture"))
     implementation(project(":ui:controller"))
     implementation(project(":ui:controller:impl"))
