+++ settings.gradle.kts
@@ -36,12 +36,16 @@ rootProject.name = "Jetpack Camera"
 include(":app")
 include(":feature:preview")
 include(":core:camera")
+include(":core:camera:testing")
 include(":core:camera:low-light")
 include(":core:camera:low-light-playservices")
 include(":feature:settings")
 include(":data:settings")
+include(":data:settings:testing")
 include(":data:media")
+include(":data:media:testing")
 include(":core:common")
+include(":core:common:testing")
 include(":benchmark")
 include(":feature:permissions")
 include(":feature:postcapture")
